Tie Location to the Window object
Per @bzbarsky and @ajklein this is where Location actually lives in
browsers. Changing this in preparation for the cross-origin objects
rewrite of Location and WindowProxy.

Has https://github.com/annevk/html-cross-origin-objects details on that
rewrite.

This fixes
https://github.com/annevk/html-cross-origin-objects/issues/24.

(We should maybe consider moving more concepts to Window. @bzbarsky
alluded to a wider mismatch between the specification and browsers.)