Add worker modules
Closes #550. The Worker and SharedWorker constructors now get an options
object, which can be specified as { type: "module", crossOrigin }, where
crossOrigin specifies the credentials mode used for the initial fetch
and for subsequent imports. The module fetching and execution machinery
is entirely reused from that for <script type="module">.

This commit does not include any modifications to importScripts; those
are in a subsequent commit.